cleaned the code and added a new module for the local storage functions with exports
